Migrate useEndpointBodies hook to redux store
