+++ workspaces/ui-v2/src/lib/Interfaces.ts
@@ -1,10 +1,9 @@
 import { ICopy } from '<src>/pages/diffs/components/ICopyRender';
-import { IRequestBody, IResponseBody } from '<src>/hooks/useEndpointBodyHook';
 import { IJsonTrail } from '../../../cli-shared/build/diffs/json-trail';
 import { IgnoreRule } from './ignore-rule';
 import { DomainIdGenerator } from './domain-id-generator';
 import { IOpticEngine } from '@useoptic/spectacle/src/index';
-import { IEndpoint } from '<src>/types';
+import { IEndpoint, IRequestBody, IResponseBody } from '<src>/types';
 
 export interface IInterpretation {
   previewTabs: IInteractionPreviewTab[];
