+++ workspaces/ui-v2/src/lib/__tests/diff-helpers/universes/makeCurrentSpecContext.ts
@@ -3,7 +3,7 @@ import {
   AllEndpointsQuery,
   endpointQueryResultsToJson,
   EndpointQueryResults,
-} from '<src>/store/endpoints';
+} from '<src>/store/endpoints/thunks';
 import { IBaseSpectacle } from '@useoptic/spectacle';
 import {
   AllRequestsAndResponsesQuery,
