+++ workspaces/ui-v2/src/pages/diffs/ReviewEndpointDiffPage/ReviewEndpointDiffPageContainer.tsx
@@ -11,7 +11,7 @@ import { useSharedDiffContext } from '<src>/pages/diffs/contexts/SharedDiffConte
 import { SpectacleContext } from '<src>/contexts/spectacle-provider';
 import { PageLayout } from '<src>/components';
 import { DiffAccessoryNavigation } from '<src>/pages/diffs/components/DiffAccessoryNavigation';
-import { useAppSelector } from '<src>/store';
+import { useAppSelector, selectors } from '<src>/store';
 
 import { ReviewEndpointDiffPage } from './ReviewEndpointDiffPage';
 
@@ -26,11 +26,7 @@ export const ReviewEndpointDiffContainer: FC<
   const spectacle = useContext(SpectacleContext)!;
 
   const endpointDiffs = useEndpointDiffs(pathId, method);
-  const endpoint = useAppSelector((state) =>
-    state.endpoints.results.data?.find(
-      (endpoint) => endpoint.pathId === pathId && endpoint.method === method
-    )
-  );
+  const endpoint = useAppSelector(selectors.getEndpoint({ pathId, method }));
   const { context } = useSharedDiffContext();
 
   const shapeDiffs = useShapeDiffInterpretations(
