+++ workspaces/ui-v2/src/pages/diffs/hooks/useAllRequestsAndResponses.ts
@@ -1,5 +1,5 @@
 import { useSpectacleQuery } from '<src>/contexts/spectacle-provider';
-import { IRequestBody, IResponseBody } from '<src>/hooks/useEndpointBodyHook';
+import { IRequestBody, IResponseBody } from '<src>/types';
 import { useMemo, useState } from 'react';
 
 export const AllRequestsAndResponsesQuery = `{
