+++ workspaces/ui-v2/src/store/endpoints/index.ts
@@ -1,0 +1,1 @@
+export * from './slice';
