+++ workspaces/ui-v2/src/types/endpoints.ts
@@ -1,4 +1,5 @@
 import { ChangelogCategory } from '<src>/hooks/useEndpointsChangelog';
+import { IChanges } from '<src>/pages/changelog/IChanges';
 
 export interface IPathParameter {
   id: string;
@@ -16,8 +17,35 @@ export interface IEndpoint {
   fullPath: string;
   pathParameters: IPathParameter[];
   isRemoved: boolean;
+  query: IQueryParameters | null;
+  requestBody: IRequestBody | null;
+  responseBodies: IResponseBody[];
 }
 
 export interface IEndpointWithChanges extends IEndpoint {
   changes: ChangelogCategory | null;
+}
+
+export interface IQueryParameters {
+  rootShapeId: string;
+  isRemoved: boolean;
+}
+
+export interface IRequestBody {
+  requestId: string;
+  contentType: string;
+  rootShapeId: string;
+  pathId: string;
+  method: string;
+  description: string;
+}
+
+export interface IResponseBody {
+  responseId: string;
+  statusCode: number;
+  contentType: string;
+  rootShapeId: string;
+  pathId: string;
+  method: string;
+  description: string;
 }
